Removed unecessary imports from build.gradle, moved ProductController to a different file
