+++ build.gradle
@@ -16,20 +16,13 @@ repositories {
 }
 
 dependencies {
-	implementation 'org.projectlombok:lombok:1.18.30'
-	annotationProcessor 'org.projectlombok:lombok:1.18.30'
-	implementation 'org.springdoc:springdoc-openapi-ui:1.6.14'
-	implementation 'org.springframework.boot:spring-boot-starter-actuator'
+	implementation 'org.projectlombok:lombok'
+	annotationProcessor 'org.projectlombok:lombok'
+	//implementation 'org.springdoc:springdoc-openapi-ui:1.6.14'
+	//implementation 'org.springframework.boot:spring-boot-starter-actuator'
 	implementation 'org.springframework.boot:spring-boot-starter-web'
-	implementation 'io.micrometer:micrometer-tracing-bridge-brave'
-	implementation 'org.projectlombok:lombok:1.18.30'
+	//implementation 'io.micrometer:micrometer-tracing-bridge-brave'
 	testImplementation 'org.springframework.boot:spring-boot-starter-test'
-
-	compileOnly 'org.projectlombok:lombok:1.18.30'
-	annotationProcessor 'org.projectlombok:lombok:1.18.30'
-
-	testCompileOnly 'org.projectlombok:lombok:1.18.30'
-	testAnnotationProcessor 'org.projectlombok:lombok:1.18.30'
 }
 
 tasks.named('test') {
